+++ CHANGELOG.md
@@ -1,6 +1,32 @@
 Change Log
 ==========
 
+## 1.4.0 (17-11-2024)
+* Update target sdk level to 34
+
+1.3.9 (06-03-2022)
+----------------------------
+*   New: Added focusDashedBorder
+
+1.3.8 (12-01-2022)
+----------------------------
+*   New: Added focusRectSizeFactor
+
+1.3.6 (24-07-2020)
+----------------------------
+*   Fix: Make focus animation step double
+
+### API Changes
+Old:
+```kotlin
+focusAnimationStep(Int)
+```
+
+New:
+```kotlin
+focusAnimationStep(Double)
+```
+
 1.3.1 (24-07-2020)
 ----------------------------
 *   Fix: Wrong focus for scaled views. [#201](https://github.com/faruktoptas/FancyShowCaseView/issues/201)